feat: rename plugins for clarify context
