+++ src/menu.ts
@@ -8,7 +8,7 @@ import { startingPages } from './providers/extracted-data';
 import promptOptions from './providers/prompt-options';
 
 // eslint-disable-next-line import/order
-import { pluginList as menuList } from 'virtual:MenuPlugins';
+import { menuPlugins as menuList } from 'virtual:MenuPlugins';
 
 import { getAvailablePluginNames } from './plugins/utils/main';
