+++ src/plugins/adblocker/main.ts
