+++ src/plugins/album-color-theme/main.ts
