+++ src/plugins/album-color-theme/renderer.ts
