+++ src/plugins/ambient-mode/renderer.ts
