+++ src/plugins/audio-compressor/renderer.ts
