+++ src/plugins/bypass-age-restrictions/renderer.ts
