+++ src/plugins/compact-sidebar/renderer.ts
