+++ src/plugins/crossfade/main.ts
