+++ src/plugins/crossfade/renderer.ts
