+++ src/plugins/disable-autoplay/renderer.ts
