+++ src/plugins/discord/main.ts
