+++ src/plugins/exponential-volume/renderer.ts
