+++ src/plugins/in-app-menu/main.ts
