+++ src/plugins/picture-in-picture/renderer.ts
