+++ src/plugins/playback-speed/renderer.ts
