+++ src/plugins/shortcuts/main.ts
