+++ src/plugins/skip-silences/renderer.ts
