+++ src/plugins/taskbar-mediacontrol/main.ts
