+++ src/plugins/visualizer/renderer.ts
