+++ src/virtual-module.d.ts
@@ -2,7 +2,7 @@ declare module 'virtual:MainPlugins' {
   import type { BrowserWindow } from 'electron';
   import type { ConfigType } from './config/dynamic';
 
-  export const pluginList: Record<string, (win: BrowserWindow, options: ConfigType) => Promise<void>>;
+  export const mainPlugins: Record<string, (win: BrowserWindow, options: ConfigType) => Promise<void>>;
 }
 
 declare module 'virtual:MenuPlugins' {
@@ -10,17 +10,17 @@ declare module 'virtual:MenuPlugins' {
   import type { MenuTemplate } from './menu';
   import type { ConfigType } from './config/dynamic';
 
-  export const pluginList: Record<string, (win: BrowserWindow, options: ConfigType, refreshMenu: () => void) => MenuTemplate>;
+  export const menuPlugins: Record<string, (win: BrowserWindow, options: ConfigType, refreshMenu: () => void) => MenuTemplate>;
 }
 
 declare module 'virtual:PreloadPlugins' {
   import type { ConfigType } from './config/dynamic';
 
-  export const pluginList: Record<string, (options: ConfigType) => Promise<void>>;
+  export const preloadPlugins: Record<string, (options: ConfigType) => Promise<void>>;
 }
 
 declare module 'virtual:RendererPlugins' {
   import type { ConfigType } from './config/dynamic';
 
-  export const pluginList: Record<string, (options: ConfigType) => Promise<void>>;
+  export const rendererPlugins: Record<string, (options: ConfigType) => Promise<void>>;
 }
